# swagger.yaml
openapi: 3.0.3
info:
  title: NexusMail API
  description: |
    ## NexusMail – Institutional Mail Management System

    A production-ready REST API for managing institutional mail with a strict 5-role hierarchy and enforced workflow transitions.

    ### Mail Workflow
    ```
    Registered → Under Review → Assigned → In Progress → Processed
    ```

    ### Roles
    | Role | Capabilities |
    |------|-------------|
    | **Admin** | Full system control |
    | **Director** | Review and assign mail |
    | **Secretary** | Register mail |
    | **Professor/Staff** | Track assigned mail |
    | **Service Lead** | Manage department workflow |

    ### Authentication
    All protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: NexusMail Support
    email: support@nexusmail.com
servers:
  - url: http://localhost:5000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { type: array }
        meta:
          type: object
          properties:
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            totalPages: { type: integer }
...
security:
  - BearerAuth: []
paths:
...
  /mails:
    get:
      tags: [Mail]
      summary: List mails with filters and pagination
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: status
          schema:
            type: string
            enum: [Registered, Under Review, Assigned, In Progress, Processed]
        - in: query
          name: type
          schema:
            type: string
            enum: [Incoming, Outgoing, Internal]
        - in: query
          name: priority
          schema:
            type: string
            enum: [Low, Medium, High, Urgent]
        - in: query
          name: isOverdue
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string, description: "Search in subject, sender, reference number" }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, updatedAt, priority, slaDeadline]
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated mail list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedResponse' }

    post:
      tags: [Mail]
      summary: Register new mail (Secretary/Admin only)
      description: |
        Creates a new mail entry. On creation, the AI pipeline runs automatically:
        - Generates an AI summary
        - Suggests a department with a confidence score
        - Calculates the SLA deadline based on category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, sender, type]
              properties:
                subject:
                  type: string
                  example: "Request for Faculty Training Budget"
                sender:
                  type: string
                  example: "Ministry of Higher Education"
                type:
                  type: string
                  enum: [Incoming, Outgoing, Internal]
                  example: "Incoming"
                category:
                  type: string
                  description: "MailCategory ObjectId"
                priority:
                  type: string
                  enum: [Low, Medium, High, Urgent]
                  default: Medium
                description:
                  type: string
                pdfUrl:
                  type: string
                  format: uri
      responses:
        '201':
          description: Mail registered with AI metadata
...
  /mails/{id}:
    get:
      tags: [Mail]
      summary: Get mail by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Mail details
        '404':
          description: Not found
  /mails/{id}/status:
    put:
      tags: [Mail]
      summary: Update mail status (validated workflow transitions)
      description: |
        **Strict transition rules:**
        - `Registered` → `Under Review` (Director only)
        - `Under Review` → `Assigned` (Director only, use /assign endpoint)
        - `Assigned` → `In Progress` (Assigned user, Director, or Admin)
        - `In Progress` → `Processed` (Assigned user, Director, or Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Registered, Under Review, Assigned, In Progress, Processed]
                note:
                  type: string
                  example: "Reviewed and proceeding with assignment"
      responses:
        '200':
          description: Status updated
        '400':
          description: Invalid status transition
  /mails/{id}/assign:
    put:
      tags: [Mail]
      summary: Assign mail to a user (Director/Admin only)
      description: |
        Mail MUST be in `Under Review` status. Assignment includes mandatory instructions.
        This automatically transitions the status to `Assigned`.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignedTo, instructions]
              properties:
                assignedTo:
                  type: string
                  description: User ObjectId to assign to
                instructions:
                  type: string
                  minLength: 5
                  example: "Please review this and prepare a technical report within 5 working days."
                assignedDepartment:
                  type: string
                  description: Department ObjectId
                priority:
                  type: string
                  enum: [Low, Medium, High, Urgent]
      responses:
        '200':
          description: Mail assigned successfully
        '400':
          description: Mail not in Under Review status
  /mails/user/{id}:
    get:
      tags: [Mail]
      summary: Get all mails for a specific user (Admin/Director/Service Lead)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: User mails

  # ── COMMENTS ─────────────────────────────────────────────────────────────────
  /mails/{id}/comments:
    post:
      tags: [Comments]
      summary: Add comment to a mail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "I will need additional time to complete the report."
                isInternal:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Comment added

    get:
      tags: [Comments]
      summary: Get all comments for a mail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Comment list

  # ── MAIL TYPES ───────────────────────────────────────────────────────────────
...